refactor: 도메인, 뷰, 메세지 별로 상수 파일 분리
