+++ src/js/__test__/vendingMachine.test.ts
@@ -1,5 +1,5 @@
 import VendingMachine from '../vendingMachine/vendingMachine';
-import { COINS } from '../constants/constants';
+import { COINS } from '../constants/vendingMachineConstants';
 
 describe('자판기 테스트', () => {
   describe('상품 관리', () => {
