+++ src/js/constants/appContants.ts
@@ -1,0 +1,14 @@
+export const URL = {
+  BASE_URL: '/javascript-vendingmachine',
+  MANAGE_ITEM: `manageItem`,
+  CHARGE_MONEY: `chargeMoney`,
+  PURCHASE_ITEM: `purchaseItem`,
+};
+
+export const CUSTOM_EVENT = {
+  ROUTE_CHANGE: 'ROUTE_CHANGE',
+  CHARGE_MONEY: 'CHARGE_MONEY',
+  ADD_ITEM: 'ADD_ITEM',
+  TABLE_ITEM_CHANGE: 'TABLE_ITEM_CHANGE',
+  TABLE_ITEM_DELETE: 'TABLE_ITEM_DELETE',
+};
