+++ src/js/constants/confirmConstants.ts
@@ -1,0 +1,3 @@
+export const CONFIRM_MESSAGE = {
+  DELETE: '정말로 삭제하시겠습니까?',
+};
