+++ src/js/constants/vendingMachineConstants.ts
@@ -1,0 +1,27 @@
+export const ITEM = {
+  NAME: {
+    MAX_LENGTH: 10,
+  },
+  PRICE: {
+    MIN: 100,
+    MAX: 10000,
+    UNIT: 10,
+  },
+  QUANTITY: {
+    MIN: 0,
+    MAX: 20,
+  },
+};
+
+export const COINS = {
+  fiveHundred: 500,
+  hundred: 100,
+  fifty: 50,
+  ten: 10,
+};
+
+export const MONEY = {
+  MIN: 0,
+  MAX: 100000,
+  UNIT: 10,
+};
