+++ src/js/constants/viewConstants.ts
@@ -1,0 +1,40 @@
+export const SELECTOR = {
+  ID: {
+    APP: '#app',
+    CONTENT: '#content',
+    CHARGE_MONEY_FORM: '#charge-money-form',
+    CURRENT_MONEY: '#current-money',
+    ADD_ITEM_NAME: '#add-item-name',
+    ADD_ITEM_PRICE: '#add-item-price',
+    ADD_ITEM_QUANTITY: '#add-item-quantity',
+    ADD_ITEM_FORM: '#add-item-form',
+  },
+  CLASS: {
+    CHARGE_MONEY_INPUT: '.charge-money-input',
+    COIN_TABLE: '.coin-table',
+    NAV_CONTAINER: '.nav-container',
+    NAV_BUTTON: '.nav-button',
+    ITEM_TABLE_BUTTON_CONTAINER: '.item-table-button-container',
+    TABLE_CONTAINER: '.table-container',
+    ITEM_TABLE_CONFIRM_BUTTON: '.item-table-confirm-button',
+    ITEM_TABLE_DELETE_BUTTON: '.item-table-delete-button',
+    ITEM_TABLE_CHANGE_BUTTON: '.item-table-change-button',
+  },
+  ID_STRING: {
+    ITEM_MANGE_TAB: 'item-manage-tab',
+    MONEY_CHARGE_TAB: 'money-charge-tab',
+    ITEM_PURCHASE_TAB: 'item-purchase-tab',
+  },
+  CLASS_STRING: {
+    NAV_BUTTON_CLICKED: 'nav-button-clicked',
+    TABLE_ITEM_NAME: 'table-item-name',
+    TABLE_ITEM_PRICE: 'table-item-price',
+    TABLE_ITEM_QUANTITY: 'table-item-quantity',
+    TABLE_ITEM_INPUT_NAME: 'table-item-input-name',
+    TABLE_ITEM_INPUT_PRICE: 'table-item-input-price',
+    TABLE_ITEM_INPUT_QUANTITY: 'table-item-input-quantity',
+    ITEM_TABLE_CHANGE_BUTTON: 'item-table-change-button',
+    ITEM_TABLE_DELETE_BUTTON: 'item-table-delete-button',
+    ITEM_TABLE_CONFIRM_BUTTON: 'item-table-confirm-button',
+  },
+};
