+++ src/js/views/chargeMoneyView.ts
@@ -2,7 +2,8 @@ import { $ } from '../utils/common';
 import { chargeMoneyTemplate, sectionTemplate } from '../templates/chareMoneyTemplate';
 import { validateInputMoney } from '../validates/validates';
 import { CoinsType } from '../types/types';
-import { SELECTOR, CUSTOM_EVENT } from '../constants/constants';
+import { CUSTOM_EVENT } from '../constants/appContants';
+import { SELECTOR } from '../constants/viewConstants';
 
 export default class ChargeMoneyView {
   $content: HTMLDivElement;
